[FIXES BUILD] Run tests only against Solidus master
When SolidusFrontend was part of Solidus, its version was lockstep
with the other components of Solidus.  As such, during this fork,
there's no need to make SolidusFrontend 3.2.0.alpha compatible with
older versions of Solidus. However, future versions of SolidusFrontend
would be expected to provide backward compatibility to supported
versions of Solidus.