Ask whether to add solidus_bolt during installation
When users answer yes, it will add solidus_bolt [1] to the Gemfile and
run its install generator. It can also be added unattended through the
`--auto-accept` option (akin to the one in the Solidus Core installer).

The prompt can be avoided altogether through the environment variable
`SKIP_SOLIDUS_BOLT`. That's applied to the CI so that it doesn't wait
forever.

[1] - https://github.com/solidusio/solidus_bolt