Skip trying to install solidus_bolt when there's no Gemfile
The dummy application generated within the extensions test suite lacks
the Gemfile [1]. We want to skip trying to install solidus_bolt in those
cases, as otherwise, it fails.

We keep the SKIP_SOLIDUS_BOLT environment variable behavior as that's
still useful to skip bolt when there's a Gemfile (i.e., when creating a
sandbox app). However, it can be removed from CI.

1 - https://github.com/solidusio/solidus/blob/bdebdbc15457487ab412702a02980bbbe32d3a6e/core/lib/generators/spree/dummy/dummy_generator.rb#L36